Add some system debugging

+++ src/mavsdk/core/mavlink_command_receiver.cpp
@@ -47,7 +47,7 @@ void MavlinkCommandReceiver::receive_command_int(const mavlink_message_t& messag
         cmd.target_component_id != MAV_COMP_ID_ALL) {
         if (_debugging) {
             LogDebug() << "Ignored command int to component " << (int)cmd.target_component_id
-                       << " instead of " << _server_component_impl.get_own_component_id();
+                       << " instead of " << (int)_server_component_impl.get_own_component_id();
         }
         return;
     }
@@ -96,7 +96,7 @@ void MavlinkCommandReceiver::receive_command_long(const mavlink_message_t& messa
         cmd.target_component_id != MAV_COMP_ID_ALL) {
         if (_debugging) {
             LogDebug() << "Ignored command long to component " << (int)cmd.target_component_id
-                       << " instead of " << _server_component_impl.get_own_component_id();
+                       << " instead of " << (int)_server_component_impl.get_own_component_id();
         }
         return;
     }
